[Feat] Edit implementation related to Measurement
- Specification changed
  - checkbox added
  - value type changed (Long to Double)
  - layout changed

+++ domain/src/main/java/com/sork/domain/entity/Measurement.kt
@@ -2,5 +2,5 @@ package com.sork.domain.entity
 
 data class Measurement(
     val type: MeasurementType,
-    val value: Long
+    val value: Double
 )
